:sailboat: Refactor mapper to specific formatting function
